@@ -170,7 +170,7 @@ impl SettableTableList<'_> for Menu {
         }
       },
       Some(Err(e)) => {
-        log::error!("{}", e);
+        log::error!("{e}");
       },
       None => {},
     }
@@ -346,7 +346,7 @@ impl Component for Menu {
       match i {
         x if x == self.schema_index => {
           let block = Block::default()
-            .title(format!(" 󰦄  {} <alt+1> (schema) ", k))
+            .title(format!(" 󰦄  {k} <alt+1> (schema) "))
             .borders(Borders::ALL)
             .border_style(if focused && self.menu_focus == MenuFocus::Schema {
               Style::default().fg(Color::Green)
