@@ -11,8 +11,8 @@ use color_eyre::eyre::{self, Result};
 use futures::stream::StreamExt;
 use sqlparser::ast::Statement;
 use sqlx::{
-  mysql::{MySql, MySqlConnectOptions, MySqlPoolOptions},
   Column, Either, Row, ValueRef,
+  mysql::{MySql, MySqlConnectOptions, MySqlPoolOptions},
 };
 use tokio::task::JoinHandle;
 
@@ -62,15 +62,16 @@ impl Database for MySqlDriver<'_> {
   }
 
   fn abort_query(&mut self) -> Result<bool> {
-    if let Some(task) = self.task.take() {
-      match task {
-        MySqlTask::Query(handle) => handle.abort(),
-        MySqlTask::TxStart(handle) => handle.abort(),
-        _ => {},
-      };
-      Ok(true)
-    } else {
-      Ok(false)
+    match self.task.take() {
+      Some(task) => {
+        match task {
+          MySqlTask::Query(handle) => handle.abort(),
+          MySqlTask::TxStart(handle) => handle.abort(),
+          _ => {},
+        };
+        Ok(true)
+      },
+      _ => Ok(false),
     }
   }
 
@@ -113,7 +114,7 @@ impl Database for MySqlDriver<'_> {
       let (results, tx) = query_with_tx(tx, &first_query).await;
       match results {
         Ok(Either::Left(rows_affected)) => {
-          log::info!("{:?} rows affected", rows_affected);
+          log::info!("{rows_affected:?} rows affected");
           (
             QueryResultsWithMetadata {
               results: Ok(Rows { headers: vec![], rows: vec![], rows_affected: Some(rows_affected) }),
@@ -138,11 +139,14 @@ impl Database for MySqlDriver<'_> {
   async fn commit_tx(&mut self) -> Result<Option<QueryResultsWithMetadata>> {
     if !matches!(self.task, Some(MySqlTask::TxPending(_))) {
       Ok(None)
-    } else if let Some(MySqlTask::TxPending(b)) = self.task.take() {
-      b.0.commit().await?;
-      Ok(Some(b.1))
     } else {
-      Ok(None)
+      match self.task.take() {
+        Some(MySqlTask::TxPending(b)) => {
+          b.0.commit().await?;
+          Ok(Some(b.1))
+        },
+        _ => Ok(None),
+      }
     }
   }
 
@@ -166,16 +170,15 @@ impl Database for MySqlDriver<'_> {
   }
 
   fn preview_rows_query(&self, schema: &str, table: &str) -> String {
-    format!("select * from `{}`.`{}` limit 100", schema, table)
+    format!("select * from `{schema}`.`{table}` limit 100")
   }
 
   fn preview_columns_query(&self, schema: &str, table: &str) -> String {
     format!(
       "select column_name, data_type, is_nullable, column_default, extra, column_comment
         from information_schema.columns
-        where table_schema = '{}' and table_name = '{}'
-        order by ordinal_position",
-      schema, table
+        where table_schema = '{schema}' and table_name = '{table}'
+        order by ordinal_position"
     )
   }
 
@@ -185,10 +188,9 @@ impl Database for MySqlDriver<'_> {
         group_concat(column_name order by ordinal_position) as column_names
         from information_schema.table_constraints
         join information_schema.key_column_usage using (constraint_schema, constraint_name, table_schema, table_name)
-        where table_schema = '{}' and table_name = '{}'
+        where table_schema = '{schema}' and table_name = '{table}'
         group by constraint_name, constraint_type, enforced
-        order by constraint_type, constraint_name",
-      schema, table
+        order by constraint_type, constraint_name"
     )
   }
 
@@ -196,9 +198,8 @@ impl Database for MySqlDriver<'_> {
     format!(
       "select index_name, column_name, non_unique, seq_in_index, index_type
         from information_schema.statistics
-        where table_schema = '{}' and table_name = '{}'
-        order by index_name, seq_in_index",
-      schema, table
+        where table_schema = '{schema}' and table_name = '{table}'
+        order by index_name, seq_in_index"
     )
   }
 
@@ -484,7 +485,7 @@ mod tests {
   use sqlparser::{ast::Statement, dialect::MySqlDialect, parser::ParserError};
 
   use super::*;
-  use crate::database::{get_execution_type, get_first_query, ExecutionType, ParseError};
+  use crate::database::{ExecutionType, ParseError, get_execution_type, get_first_query};
 
   #[test]
   fn test_get_first_query() {
@@ -561,7 +562,7 @@ mod tests {
         (Err(ParseError::SqlParserError(msg)), Err(ParseError::SqlParserError(expected_msg))) => {
           assert_eq!(msg, expected_msg)
         },
-        _ => panic!("Unexpected result for input: {}", input),
+        _ => panic!("Unexpected result for input: {input}"),
       }
     }
   }
@@ -584,8 +585,7 @@ mod tests {
       assert_eq!(
         get_execution_type(query.to_string(), false, Driver::MySql).unwrap().0,
         expected,
-        "Failed for query: {}",
-        query
+        "Failed for query: {query}"
       );
     }
   }
