@@ -11,9 +11,9 @@ use color_eyre::eyre::{self, Result};
 use futures::stream::StreamExt;
 use sqlparser::ast::Statement;
 use sqlx::{
+  Column, Either, Row, ValueRef,
   sqlite::{Sqlite, SqliteConnectOptions, SqlitePoolOptions},
   types::uuid,
-  Column, Either, Row, ValueRef,
 };
 
 use super::{Database, DbTaskResult, Driver, Header, Headers, QueryResultsWithMetadata, QueryTask, Rows, Value};
@@ -62,15 +62,16 @@ impl Database for SqliteDriver<'_> {
   }
 
   fn abort_query(&mut self) -> Result<bool> {
-    if let Some(task) = self.task.take() {
-      match task {
-        SqliteTask::Query(handle) => handle.abort(),
-        SqliteTask::TxStart(handle) => handle.abort(),
-        _ => {},
-      };
-      Ok(true)
-    } else {
-      Ok(false)
+    match self.task.take() {
+      Some(task) => {
+        match task {
+          SqliteTask::Query(handle) => handle.abort(),
+          SqliteTask::TxStart(handle) => handle.abort(),
+          _ => {},
+        };
+        Ok(true)
+      },
+      _ => Ok(false),
     }
   }
 
@@ -113,7 +114,7 @@ impl Database for SqliteDriver<'_> {
       let (results, tx) = query_with_tx(tx, &first_query).await;
       match results {
         Ok(Either::Left(rows_affected)) => {
-          log::info!("{:?} rows affected", rows_affected);
+          log::info!("{rows_affected:?} rows affected");
           (
             QueryResultsWithMetadata {
               results: Ok(Rows { headers: vec![], rows: vec![], rows_affected: Some(rows_affected) }),
@@ -138,11 +139,14 @@ impl Database for SqliteDriver<'_> {
   async fn commit_tx(&mut self) -> Result<Option<QueryResultsWithMetadata>> {
     if !matches!(self.task, Some(SqliteTask::TxPending(_))) {
       Ok(None)
-    } else if let Some(SqliteTask::TxPending(b)) = self.task.take() {
-      b.0.commit().await?;
-      Ok(Some(b.1))
     } else {
-      Ok(None)
+      match self.task.take() {
+        Some(SqliteTask::TxPending(b)) => {
+          b.0.commit().await?;
+          Ok(Some(b.1))
+        },
+        _ => Ok(None),
+      }
     }
   }
 
@@ -167,19 +171,19 @@ impl Database for SqliteDriver<'_> {
   }
 
   fn preview_rows_query(&self, schema: &str, table: &str) -> String {
-    format!("select * from \"{}\" limit 100", table)
+    format!("select * from \"{table}\" limit 100")
   }
 
   fn preview_columns_query(&self, schema: &str, table: &str) -> String {
-    format!("pragma table_info(\"{}\")", table)
+    format!("pragma table_info(\"{table}\")")
   }
 
   fn preview_constraints_query(&self, schema: &str, table: &str) -> String {
-    format!("pragma foreign_key_list(\"{}\")", table)
+    format!("pragma foreign_key_list(\"{table}\")")
   }
 
   fn preview_indexes_query(&self, schema: &str, table: &str) -> String {
-    format!("pragma index_list(\"{}\")", table)
+    format!("pragma index_list(\"{table}\")")
   }
 
   fn preview_policies_query(&self, schema: &str, table: &str) -> String {
@@ -313,22 +317,27 @@ fn parse_value(row: &<Sqlite as sqlx::Database>::Row, col: &<Sqlite as sqlx::Dat
     )),
     "TEXT" => {
       // Try parsing as different types that might be stored as TEXT
-      if let Ok(dt) = row.try_get::<chrono::NaiveDateTime, _>(col.ordinal()) {
-        Some(Value { parse_error: false, string: dt.to_string(), is_null: false })
-      } else if let Ok(dt) = row.try_get::<chrono::DateTime<chrono::Utc>, _>(col.ordinal()) {
-        Some(Value { parse_error: false, string: dt.to_string(), is_null: false })
-      } else if let Ok(date) = row.try_get::<chrono::NaiveDate, _>(col.ordinal()) {
-        Some(Value { parse_error: false, string: date.to_string(), is_null: false })
-      } else if let Ok(time) = row.try_get::<chrono::NaiveTime, _>(col.ordinal()) {
-        Some(Value { parse_error: false, string: time.to_string(), is_null: false })
-      } else if let Ok(uuid) = row.try_get::<uuid::Uuid, _>(col.ordinal()) {
-        Some(Value { parse_error: false, string: uuid.to_string(), is_null: false })
-      } else if let Ok(json) = row.try_get::<serde_json::Value, _>(col.ordinal()) {
-        Some(Value { parse_error: false, string: json.to_string(), is_null: false })
-      } else if let Ok(string) = row.try_get::<String, _>(col.ordinal()) {
-        Some(Value { parse_error: false, string, is_null: false })
-      } else {
-        Some(Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false })
+      match row.try_get::<chrono::NaiveDateTime, _>(col.ordinal()) {
+        Ok(dt) => Some(Value { parse_error: false, string: dt.to_string(), is_null: false }),
+        _ => match row.try_get::<chrono::DateTime<chrono::Utc>, _>(col.ordinal()) {
+          Ok(dt) => Some(Value { parse_error: false, string: dt.to_string(), is_null: false }),
+          _ => match row.try_get::<chrono::NaiveDate, _>(col.ordinal()) {
+            Ok(date) => Some(Value { parse_error: false, string: date.to_string(), is_null: false }),
+            _ => match row.try_get::<chrono::NaiveTime, _>(col.ordinal()) {
+              Ok(time) => Some(Value { parse_error: false, string: time.to_string(), is_null: false }),
+              _ => match row.try_get::<uuid::Uuid, _>(col.ordinal()) {
+                Ok(uuid) => Some(Value { parse_error: false, string: uuid.to_string(), is_null: false }),
+                _ => match row.try_get::<serde_json::Value, _>(col.ordinal()) {
+                  Ok(json) => Some(Value { parse_error: false, string: json.to_string(), is_null: false }),
+                  _ => match row.try_get::<String, _>(col.ordinal()) {
+                    Ok(string) => Some(Value { parse_error: false, string, is_null: false }),
+                    _ => Some(Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false }),
+                  },
+                },
+              },
+            },
+          },
+        },
       }
     },
     "BLOB" => Some(row.try_get::<Vec<u8>, usize>(col.ordinal()).map_or(
@@ -350,41 +359,36 @@ fn parse_value(row: &<Sqlite as sqlx::Database>::Row, col: &<Sqlite as sqlx::Dat
     )),
     "DATETIME" => {
       // Similar to TEXT, but we'll try timestamp first
-      if let Ok(dt) = row.try_get::<i64, _>(col.ordinal()) {
-        Some(chrono::DateTime::from_timestamp(dt, 0).map_or(
+      match row.try_get::<i64, _>(col.ordinal()) {
+        Ok(dt) => Some(chrono::DateTime::from_timestamp(dt, 0).map_or(
           Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false },
           |received| Value { parse_error: false, string: received.to_string(), is_null: false },
-        ))
-      } else if let Ok(dt) = row.try_get::<chrono::NaiveDateTime, _>(col.ordinal()) {
-        Some(Value { parse_error: true, string: dt.to_string(), is_null: false })
-      } else if let Ok(dt) = row.try_get::<chrono::DateTime<chrono::Utc>, _>(col.ordinal()) {
-        Some(Value { parse_error: true, string: dt.to_string(), is_null: false })
-      } else {
-        Some(row.try_get::<String, usize>(col.ordinal()).map_or(
-          Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false },
-          |received| Value { parse_error: false, string: received.to_string(), is_null: false },
-        ))
+        )),
+        _ => match row.try_get::<chrono::NaiveDateTime, _>(col.ordinal()) {
+          Ok(dt) => Some(Value { parse_error: true, string: dt.to_string(), is_null: false }),
+          _ => match row.try_get::<chrono::DateTime<chrono::Utc>, _>(col.ordinal()) {
+            Ok(dt) => Some(Value { parse_error: true, string: dt.to_string(), is_null: false }),
+            _ => Some(row.try_get::<String, usize>(col.ordinal()).map_or(
+              Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false },
+              |received| Value { parse_error: false, string: received.to_string(), is_null: false },
+            )),
+          },
+        },
       }
     },
-    "DATE" => {
-      if let Ok(date) = row.try_get::<chrono::NaiveDate, _>(col.ordinal()) {
-        Some(Value { parse_error: true, string: date.to_string(), is_null: false })
-      } else {
-        Some(row.try_get::<String, usize>(col.ordinal()).map_or(
-          Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false },
-          |received| Value { parse_error: false, string: received.to_string(), is_null: false },
-        ))
-      }
+    "DATE" => match row.try_get::<chrono::NaiveDate, _>(col.ordinal()) {
+      Ok(date) => Some(Value { parse_error: true, string: date.to_string(), is_null: false }),
+      _ => Some(row.try_get::<String, usize>(col.ordinal()).map_or(
+        Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false },
+        |received| Value { parse_error: false, string: received.to_string(), is_null: false },
+      )),
     },
-    "TIME" => {
-      if let Ok(time) = row.try_get::<chrono::NaiveTime, _>(col.ordinal()) {
-        Some(Value { parse_error: true, string: time.to_string(), is_null: false })
-      } else {
-        Some(row.try_get::<String, usize>(col.ordinal()).map_or(
-          Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false },
-          |received| Value { parse_error: false, string: received.to_string(), is_null: false },
-        ))
-      }
+    "TIME" => match row.try_get::<chrono::NaiveTime, _>(col.ordinal()) {
+      Ok(time) => Some(Value { parse_error: true, string: time.to_string(), is_null: false }),
+      _ => Some(row.try_get::<String, usize>(col.ordinal()).map_or(
+        Value { parse_error: true, string: "_ERROR_".to_string(), is_null: false },
+        |received| Value { parse_error: false, string: received.to_string(), is_null: false },
+      )),
     },
     _ => {
       // For any other types, try to cast to string
@@ -401,7 +405,7 @@ mod tests {
   use sqlparser::{ast::Statement, dialect::SQLiteDialect, parser::ParserError};
 
   use super::*;
-  use crate::database::{get_execution_type, get_first_query, ExecutionType, ParseError};
+  use crate::database::{ExecutionType, ParseError, get_execution_type, get_first_query};
 
   #[test]
   fn test_get_first_query() {
@@ -478,7 +482,7 @@ mod tests {
         (Err(ParseError::SqlParserError(msg)), Err(ParseError::SqlParserError(expected_msg))) => {
           assert_eq!(msg, expected_msg)
         },
-        _ => panic!("Unexpected result for input: {}", input),
+        _ => panic!("Unexpected result for input: {input}"),
       }
     }
   }
@@ -501,8 +505,7 @@ mod tests {
       assert_eq!(
         get_execution_type(query.to_string(), false, Driver::Sqlite).unwrap().0,
         expected,
-        "Failed for query: {}",
-        query
+        "Failed for query: {query}"
       );
     }
   }
