@@ -23,7 +23,7 @@ pub fn get_password(connection_name: &str, username: &str) -> Result<Password> {
     },
     Err(e) => match e {
       keyring::Error::NoEntry => {
-        println!("{}@{}", username, connection_name);
+        println!("{username}@{connection_name}");
         let password = rpassword::prompt_password("Password: ")?;
 
         print!("Save password (Y/n): ");
