@@ -21,7 +21,7 @@ use std::{
 };
 
 use clap::Parser;
-use cli::{extract_driver_from_url, prompt_for_database_selection, Cli, Driver};
+use cli::{Cli, Driver, extract_driver_from_url, prompt_for_database_selection};
 use color_eyre::eyre::Result;
 use config::{Config, ConnectionString};
 use dotenvy::dotenv;
@@ -104,11 +104,12 @@ async fn tokio_main() -> Result<()> {
 
 #[tokio::main]
 async fn main() -> Result<()> {
-  if let Err(e) = tokio_main().await {
-    eprintln!("{} error: Something went wrong", env!("CARGO_PKG_NAME"));
-    Err(e)
-  } else {
-    Ok(())
+  match tokio_main().await {
+    Err(e) => {
+      eprintln!("{} error: Something went wrong", env!("CARGO_PKG_NAME"));
+      Err(e)
+    },
+    _ => Ok(()),
   }
 }
 
